binance_data/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-binance
python-dotenv
